requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bleach==5.0.1
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-ipware==4.0.2
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-api-logger==1.1.10
drf-spectacular==0.24.0
et-xmlfile==1.1.0
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.16.0
numpy==1.23.3
openpyxl==3.0.10
pandas==1.4.4
pkgutil-resolve-name==1.3.10
PyJWT==2.4.0
PyMySQL==1.0.2
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
webencodings==0.5.1
zipp==3.8.1
